social_media_api/requirements.txt
Django==5.1.5
django-filter==25.1
djangorestframework==3.15.2
Pillow>=9.5.0



